de-05-homework/zoomcamp/pipeline/assets/ingestion/requirements.txt
pandas
requests
pyarrow
python-dateutil
pytzdata
